requirements copy.txt
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
awsebcli==3.18.2
billiard==3.6.3.0
boto3==1.14.21
botocore==1.15.49
cachetools==4.1.1
celery==4.4.7
cement==2.8.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-rest-framework==0.1.0
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
future==0.16.0
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.7
importlib-metadata==1.7.0
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pathspec==0.5.9
pika==1.1.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.5.3
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2020.1
pywin32==228
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
semantic-version==2.5.0
simplejson==3.17.2
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.9
wrapt==1.12.1
zipp==3.1.0
